Lesson 5.5

+++ source/Lesson5.4/PointLightEffect.cpp
@@ -65,7 +65,7 @@ namespace Rendering
 		const vector<vec3>& normals = mesh.Normals();
 		assert(normals.size() == sourceVertices.size());
 
-		for (UINT i = 0; i < sourceVertices.size(); i++)
+		for (size_t i = 0; i < sourceVertices.size(); i++)
 		{
 			const vec3& position = sourceVertices.at(i);
 			const vec2& uv = static_cast<vec2>(textureCoordinates->at(i));
